+++ com.unity.render-pipelines.universal/Editor/ShaderPreprocessor.cs
@@ -725,13 +725,8 @@ namespace UnityEditor.Rendering.Universal
                         ++i;
                 }
 
-                if (compilerDataList is List<ShaderCompilerData> inputDataList)
-                    inputDataList.RemoveRange(inputShaderVariantCount, inputDataList.Count - inputShaderVariantCount);
-                else
-                {
-                    for (int i = compilerDataList.Count - 1; i >= inputShaderVariantCount; --i)
-                        compilerDataList.RemoveAt(i);
-                }
+                compilerDataList.RemoveElementsInRange(inputShaderVariantCount, compilerDataList.Count - inputShaderVariantCount);
+
             }
 
             LogShaderVariants(shader, snippetData, prevVariantCount, compilerDataList.Count, stripTimeMs);