+++ com.unity.render-pipelines.core/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Linear version of function that sets FSR RCAS shader constants
 - `DebugUI.ObjectPopupField` to render a list of `UnityEngine.Objects` as a popup on the Rendering Debugger.
 - Add probe volume influence weight parameter
+- Added a new extension for `IList` to remove a given number of elements from the back of it.
 - Added support for multiple Baking States to Prove Volumes.
 - Hidding Volume Components not available for the current pipeline on the Volume Profile Inspector.
 - Added error on ResourceReloader when attempting to use [ReloadGroup] on ScriptableObject.